+++ example-dogs/src/main/java/com/zhuinden/simplestackcomposedogexample/app/MainActivity.kt
@@ -10,24 +10,25 @@ import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
 import androidx.compose.ui.Modifier
 import com.zhuinden.simplestack.AsyncStateChanger
+import com.zhuinden.simplestack.BackHandlingModel
+import com.zhuinden.simplestack.Backstack
 import com.zhuinden.simplestack.History
 import com.zhuinden.simplestack.navigator.Navigator
 import com.zhuinden.simplestackcomposedogexample.features.doglist.DogListKey
 import com.zhuinden.simplestackcomposeintegration.core.BackstackProvider
 import com.zhuinden.simplestackcomposeintegration.core.ComposeStateChanger
+import com.zhuinden.simplestackextensions.lifecyclektx.observeAheadOfTimeWillHandleBackChanged
 import com.zhuinden.simplestackextensions.navigatorktx.androidContentFrame
 import com.zhuinden.simplestackextensions.services.DefaultServiceProvider
 
 class MainActivity : AppCompatActivity() {
     private val composeStateChanger = ComposeStateChanger()
 
-    private val backPressedCallback = object: OnBackPressedCallback(true) {
+    private lateinit var backstack: Backstack
+
+    private val backPressedCallback = object : OnBackPressedCallback(false) {
         override fun handleOnBackPressed() {
-            if (!Navigator.onBackPressed(this@MainActivity)) {
-                this.remove() // this is the only safe way to invoke onBackPressed while cancelling the execution of this callback
-                onBackPressed() // this is the only safe way to invoke onBackPressed while cancelling the execution of this callback
-                this@MainActivity.onBackPressedDispatcher.addCallback(this) // this is the only safe way to invoke onBackPressed while cancelling the execution of this callback
-            }
+            backstack.goBack()
         }
     }
 
@@ -38,7 +39,8 @@ class MainActivity : AppCompatActivity() {
 
         val app = application as CustomApplication
 
-        val backstack = Navigator.configure()
+        backstack = Navigator.configure()
+            .setBackHandlingModel(BackHandlingModel.AHEAD_OF_TIME)
             .setGlobalServices(app.globalServices)
             .setScopedServices(DefaultServiceProvider())
             .setStateChanger(AsyncStateChanger(composeStateChanger))
@@ -55,11 +57,11 @@ class MainActivity : AppCompatActivity() {
                 }
             }
         }
-    }
 
-    @Suppress("DEPRECATION")
-    @Deprecated("Deprecated in Java")
-    override final fun onBackPressed() {
-        super.onBackPressed()
+        backPressedCallback.isEnabled = backstack.willHandleAheadOfTimeBack()
+
+        backstack.observeAheadOfTimeWillHandleBackChanged(this) {
+            backPressedCallback.isEnabled = it
+        }
     }
 }