+++ example-ftue/src/main/java/com/zhuinden/simplestackftuecomposesample/app/MainActivity.kt
@@ -9,13 +9,11 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
 import androidx.compose.ui.Modifier
-import com.zhuinden.simplestack.AsyncStateChanger
-import com.zhuinden.simplestack.History
-import com.zhuinden.simplestack.StateChange
-import com.zhuinden.simplestack.StateChanger
+import com.zhuinden.simplestack.*
 import com.zhuinden.simplestack.navigator.Navigator
 import com.zhuinden.simplestackcomposeintegration.core.BackstackProvider
 import com.zhuinden.simplestackcomposeintegration.core.ComposeStateChanger
+import com.zhuinden.simplestackextensions.lifecyclektx.observeAheadOfTimeWillHandleBackChanged
 import com.zhuinden.simplestackextensions.navigatorktx.androidContentFrame
 import com.zhuinden.simplestackextensions.servicesktx.get
 import com.zhuinden.simplestackftuecomposesample.features.login.LoginKey
@@ -26,17 +24,13 @@ class MainActivity : AppCompatActivity(), AsyncStateChanger.NavigationHandler {
     private lateinit var composeStateChanger: ComposeStateChanger
     private lateinit var authenticationManager: AuthenticationManager
 
-    @Suppress("DEPRECATION")
-    private val backPressedCallback =
-        object : OnBackPressedCallback(true) { // this is the only way to make Compose BackHandler work reliably for now
-            override fun handleOnBackPressed() {
-                if (!Navigator.onBackPressed(this@MainActivity)) {
-                    this.remove()
-                    onBackPressed()
-                    this@MainActivity.onBackPressedDispatcher.addCallback(this)
-                }
-            }
+    private lateinit var backstack: Backstack
+
+    private val backPressedCallback = object : OnBackPressedCallback(false) {
+        override fun handleOnBackPressed() {
+            backstack.goBack()
         }
+    }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -50,7 +44,8 @@ class MainActivity : AppCompatActivity(), AsyncStateChanger.NavigationHandler {
 
         authenticationManager = globalServices.get()
 
-        val backstack = Navigator.configure()
+        backstack = Navigator.configure()
+            .setBackHandlingModel(BackHandlingModel.AHEAD_OF_TIME)
             .setStateChanger(AsyncStateChanger(this))
             .setScopedServices(ServiceProvider())
             .setGlobalServices(globalServices)
@@ -62,6 +57,12 @@ class MainActivity : AppCompatActivity(), AsyncStateChanger.NavigationHandler {
                     }
                 )
             )
+
+        backPressedCallback.isEnabled = backstack.willHandleAheadOfTimeBack()
+
+        backstack.observeAheadOfTimeWillHandleBackChanged(this) {
+            backPressedCallback.isEnabled = it
+        }
 
         setContent {
             BackstackProvider(backstack) {