+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change log
 
+-Simple Stack Compose Integration 0.11.0 (2023-02-23)
+--------------------------------
+
+- UPDATE: simple-stack to 2.7.0, simple-stack-extensions 2.3.0.
+
 -Simple Stack Compose Integration 0.10.0 (2023-02-23)
 --------------------------------